added cirotix's patch as outlined below
With this patch, the options and comments settings for the nodes
created from the import of a file have the default values set in
admin/content/types/biblio and are not hard-coded anymore.

It also corrects a small bug that was preventing the citekey to be
exported (previously submitted in an other patch).