Ported the following chages from the 5.x branch...
Added $base to the argument list of the theme_biblio_long, theme_biblio_short and theme_biblio_list in order to cut down on the number of db hits to get the $base variable.

Added "default" section in the main switch in biblio_db_search().  This allows filtering on any of the biblio db fields.